ml-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
scikit-learn==1.6.0
pandas==2.2.2
numpy==2.2.6
joblib==1.4.2
pydantic>=2.0
python-dotenv>=1.0.0
boto3>=1.34.0
pytest>=8.0.0
httpx>=0.27.0
